Add files via upload
This is actually one of my earliest works I've done using Selenium in Python. I decided to upload any of my works whether it's a "simple" thing or complicated lines of codes. 

If you ask "what's that?" let me explain a bit. So lately, there's been a question asked to a contestant on the show "Who wants to be a millionaire?" and the question went like this: Which of these four letters is found less than the other three in the names of the 81 provinces in Turkey? A- ) Ş B- ) V C- ) G D- ) H and when I saw the question I asked myself why not implement it on a Python application? So that's what this code does: It get's the letter that user wants to search in the names of 81 cities in Turkey. Then it counts in how many cities does this letter occur and then prints the answer!